feat: bring this module into the 21st century
