+++ .gitignore
@@ -1,2 +1,6 @@
 .*swp
 .DS_Store
+package-lock.json
+*~
+node_modules
+lib/
